Refactored network package
