+++ app/src/main/java/com/ikvych/cocktail/data/network/base/BaseApiRepository.kt
@@ -1,0 +1,4 @@
+package com.ikvych.cocktail.data.network.base
+
+interface BaseApiRepository {
+}
