+++ app/src/main/java/com/ikvych/cocktail/data/network/base/DrinkApiRepository.kt
@@ -1,10 +1,10 @@
-package com.ikvych.cocktail.data.repository.base
+package com.ikvych.cocktail.data.network.base
 
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.ikvych.cocktail.data.db.model.Drink
+import com.ikvych.cocktail.data.repository.source.base.BaseRepository
 
-interface DrinkRepository : BaseRepository{
+interface DrinkApiRepository : BaseRepository {
 
     // Methods for work with Api
     fun getDrinkApiLiveData(): MutableLiveData<List<Drink>>
