--- app/src/main/java/com/ikvych/cocktail/data/repository/base/BaseRepository.kt
@@ -1,4 +1,0 @@
-package com.ikvych.cocktail.data.repository.base
-
-interface BaseRepository {
-}
