+++ app/src/main/java/com/ikvych/cocktail/viewmodel/SearchActivityViewModel.kt
@@ -5,29 +5,29 @@ import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.SavedStateHandle
 import com.ikvych.cocktail.data.db.model.Drink
-import com.ikvych.cocktail.data.repository.DrinkRepositoryImpl
+import com.ikvych.cocktail.data.network.DrinkApiRepositoryImpl
 import com.ikvych.cocktail.data.repository.source.DrinkRepository
 import com.ikvych.cocktail.viewmodel.base.BaseViewModel
 
 
 class SearchActivityViewModel(
-    private val drinkRepository1: DrinkRepository,
+    private val drinkRepository: DrinkRepository,
     application: Application,
     savedStateHandle: SavedStateHandle
 ) : BaseViewModel(application, savedStateHandle) {
 
-    protected val drinkRepository: com.ikvych.cocktail.data.repository.base.DrinkRepository = DrinkRepositoryImpl(application)
+    protected val drinkApiRepository: com.ikvych.cocktail.data.network.base.DrinkApiRepository = DrinkApiRepositoryImpl(application)
 
     val drinkLiveData: MutableLiveData<List<Drink>> =
         object : MediatorLiveData<List<Drink>>() {
             init {
-                addSource(drinkRepository.getDrinkApiLiveData()) {
-                    value = drinkRepository.getDrinkApiLiveData().value
+                addSource(drinkApiRepository.getDrinkApiLiveData()) {
+                    value = drinkApiRepository.getDrinkApiLiveData().value
                 }
             }
         }
 
     fun updateDrinksLiveData(query: String) {
-        drinkRepository.updateDrinkApiLiveData(query)
+        drinkApiRepository.updateDrinkApiLiveData(query)
     }
 }
